# backend/app-store/src/client/openapi.yaml
...
openapi: 3.0.3
info:
  title: app-store API
  version: 1.0-SNAPSHOT
servers:
- url: http://localhost:30001
paths:
...
  /person/getAll:
    get:
      tags:
      - Person Api
      description: get all people
      operationId: getAllPeople
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PersonDTO'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - ADMIN
      - SecurityScheme:
        - ADMIN
  /person/getById:
    get:
      tags:
      - Person Api
      description: get person
      operationId: getPerson
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  bid:
                    format: int64
                    type: integer
                  username:
                    type: string
                  name:
                    type: string
                  lastName:
                    type: string
                  addresses:
                    type: array
                    items:
                      $ref: '#/components/schemas/AddressDTO'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - USER
      - SecurityScheme:
        - USER
  /person/update:
    patch:
      tags:
      - Person Api
      description: update person
      operationId: updatePerson
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Person'
      responses:
        "202":
          description: ACCEPTED
          content:
            application/json:
              schema:
                type: object
                properties:
                  bid:
                    format: int64
                    type: integer
                  username:
                    type: string
                  name:
                    type: string
                  lastName:
                    type: string
                  addresses:
                    type: array
                    items:
                      $ref: '#/components/schemas/AddressDTO'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - USER
      - SecurityScheme:
        - USER
  /product/create:
    post:
      tags:
      - Product Api
      description: create product
      operationId: createProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        "202":
          description: ACCEPTED
          content:
            application/json:
              schema:
                type: object
                properties:
                  bid:
                    format: int64
                    type: integer
                  thumbnail:
                    type: string
                  pictures:
                    type: string
                  productCategory:
                    $ref: '#/components/schemas/ProductCategory'
                  name:
                    type: string
                  producer:
                    type: string
                  guarantee:
                    type: string
                  model:
                    type: string
                  price:
                    format: double
                    type: number
                  quantity:
                    format: int32
                    type: integer
                  rate:
                    format: float
                    type: number
                  charger:
                    $ref: '#/components/schemas/ChargerDTO'
                  cooler:
                    $ref: '#/components/schemas/CoolerDTO'
                  dramMemory:
                    $ref: '#/components/schemas/DRAMMemoryDTO'
                  graphicCard:
                    $ref: '#/components/schemas/GraphicCardDTO'
                  hardDrive:
                    $ref: '#/components/schemas/HardDriveDTO'
                  motherboard:
                    $ref: '#/components/schemas/MotherboardDTO'
                  pcCase:
                    $ref: '#/components/schemas/PcCaseDTO'
                  processor:
                    $ref: '#/components/schemas/ProcessorDTO'
                  opinions:
                    type: array
                    items:
                      $ref: '#/components/schemas/OpinionDto'
...
  /product/getById/id/{id}:
    get:
      tags:
      - Product Api
      description: get product by id
      operationId: getProductById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  bid:
                    format: int64
                    type: integer
                  thumbnail:
                    type: string
                  pictures:
                    type: string
                  productCategory:
                    $ref: '#/components/schemas/ProductCategory'
                  name:
                    type: string
                  producer:
                    type: string
                  guarantee:
                    type: string
                  model:
                    type: string
                  price:
                    format: double
                    type: number
                  quantity:
                    format: int32
                    type: integer
                  productOrders:
                    uniqueItems: true
                    type: array
                    items:
                      $ref: '#/components/schemas/ProductOrder'
                  charger:
                    $ref: '#/components/schemas/Charger'
                  cooler:
                    $ref: '#/components/schemas/Cooler'
                  dramMemory:
                    $ref: '#/components/schemas/DRAMMemory'
                  graphicCard:
                    $ref: '#/components/schemas/GraphicCard'
                  hardDrive:
                    $ref: '#/components/schemas/HardDrive'
                  motherboard:
                    $ref: '#/components/schemas/Motherboard'
                  pcCase:
                    $ref: '#/components/schemas/PcCase'
                  processor:
                    $ref: '#/components/schemas/Processor'
                  rate:
                    format: float
                    type: number
                  thumbnailAsByte:
                    type: string
                  picturesAsBytes:
                    type: string
  /product/updateBaseById/id/{id}:
    put:
      tags:
      - Product Api
      description: update product base attributes by id
      operationId: updateProductBaseById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductSearchCriteria'
      responses:
        "202":
          description: ACCEPTED
          content:
            application/json:
              schema:
                type: object
                properties:
                  bid:
                    format: int64
                    type: integer
                  thumbnail:
                    type: string
                  pictures:
                    type: string
                  productCategory:
                    $ref: '#/components/schemas/ProductCategory'
                  name:
                    type: string
                  producer:
                    type: string
                  guarantee:
                    type: string
                  model:
                    type: string
                  price:
                    format: double
                    type: number
                  quantity:
                    format: int32
                    type: integer
                  rate:
                    format: float
                    type: number
                  charger:
                    $ref: '#/components/schemas/ChargerDTO'
                  cooler:
                    $ref: '#/components/schemas/CoolerDTO'
                  dramMemory:
                    $ref: '#/components/schemas/DRAMMemoryDTO'
                  graphicCard:
                    $ref: '#/components/schemas/GraphicCardDTO'
                  hardDrive:
                    $ref: '#/components/schemas/HardDriveDTO'
                  motherboard:
                    $ref: '#/components/schemas/MotherboardDTO'
                  pcCase:
                    $ref: '#/components/schemas/PcCaseDTO'
                  processor:
                    $ref: '#/components/schemas/ProcessorDTO'
                  opinions:
                    type: array
                    items:
                      $ref: '#/components/schemas/OpinionDto'
  /product/updateProductWithDetails:
    put:
      tags:
      - Product Api
      description: update product with details by id
      operationId: updateProductWithDetailsById
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        "202":
          description: ACCEPTED
          content:
            application/json:
              schema:
                type: object
                properties:
                  bid:
                    format: int64
                    type: integer
                  thumbnail:
                    type: string
                  pictures:
                    type: string
                  productCategory:
                    $ref: '#/components/schemas/ProductCategory'
                  name:
                    type: string
                  producer:
                    type: string
                  guarantee:
                    type: string
                  model:
                    type: string
                  price:
                    format: double
                    type: number
                  quantity:
                    format: int32
                    type: integer
                  rate:
                    format: float
                    type: number
                  charger:
                    $ref: '#/components/schemas/ChargerDTO'
                  cooler:
                    $ref: '#/components/schemas/CoolerDTO'
                  dramMemory:
                    $ref: '#/components/schemas/DRAMMemoryDTO'
                  graphicCard:
                    $ref: '#/components/schemas/GraphicCardDTO'
                  hardDrive:
                    $ref: '#/components/schemas/HardDriveDTO'
                  motherboard:
                    $ref: '#/components/schemas/MotherboardDTO'
                  pcCase:
                    $ref: '#/components/schemas/PcCaseDTO'
                  processor:
                    $ref: '#/components/schemas/ProcessorDTO'
                  opinions:
                    type: array
                    items:
                      $ref: '#/components/schemas/OpinionDto'
...
components:
  schemas:
    Address:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        province:
          type: string
        zipCode:
          type: string
        streetName:
          type: string
        buildingNumber:
          type: string
        apartmentNumber:
          format: int32
          type: integer
    AddressDTO:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        province:
          type: string
        zipCode:
          type: string
        streetName:
          type: string
        buildingNumber:
          type: string
        apartmentNumber:
          format: int32
          type: integer
...
    Charger:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        power:
          type: string
        standard:
          type: string
        noise:
          format: int32
          type: integer
        coolingType:
          type: string
        width:
          format: int32
          type: integer
        height:
          format: int32
          type: integer
        depth:
          format: int32
          type: integer
    ChargerDTO:
      type: object
      properties:
        power:
          type: string
        standard:
          type: string
        coolingType:
          type: string
        noise:
          format: int32
          type: integer
        width:
          format: int32
          type: integer
        height:
          format: int32
          type: integer
        depth:
          format: int32
          type: integer
    Cooler:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        type:
          type: string
        maxRotationSpeed:
          format: int32
          type: integer
        maxVolume:
          format: int32
          type: integer
        supplyVoltage:
          format: int32
          type: integer
        coolerType:
          type: string
    CoolerDTO:
      type: object
      properties:
        type:
          type: string
        maxRotationSpeed:
          format: int32
          type: integer
        maxVolume:
          format: int32
          type: integer
        supplyVoltage:
          format: int32
          type: integer
        coolerType:
          type: string
    DRAMMemory:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        latencyCycle:
          type: string
        memory:
          format: int32
          type: integer
        memoryType:
          type: string
        frequency:
          type: string
    DRAMMemoryDTO:
      type: object
      properties:
        latencyCycle:
          type: string
        memoryType:
          type: string
        frequency:
          type: string
        memory:
          format: int32
          type: integer
...
    GraphicCard:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        memoryChipset:
          type: string
        memory:
          format: int32
          type: integer
        connectorType:
          type: string
        memoryType:
          type: string
        memoryClocking:
          type: string
    GraphicCardDTO:
      type: object
      properties:
        memoryChipset:
          type: string
        memory:
          format: int32
          type: integer
        connectorType:
          type: string
        memoryType:
          type: string
        memoryClocking:
          type: string
    HardDrive:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        memoryInterface:
          type: string
        memory:
          format: int32
          type: integer
        memoryType:
          type: string
    HardDriveDTO:
      type: object
      properties:
        memoryInterface:
          type: string
        memory:
          format: int32
          type: integer
        memoryType:
          type: string
    Motherboard:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        motherboardStandard:
          type: string
        processorSocket:
          type: string
        memoryType:
          type: string
        maxMemory:
          format: int32
          type: integer
    MotherboardDTO:
      type: object
      properties:
        motherboardStandard:
          type: string
        processorSocket:
          type: string
        memoryType:
          type: string
        maxMemory:
          format: int32
          type: integer
    OpinionDto:
      type: object
      properties:
        comment:
          type: string
        rate:
          format: float
          type: number
        username:
          type: string
        productId:
          format: int64
          type: integer
    PcCase:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        width:
          format: double
          type: number
        length:
          format: double
          type: number
        depth:
          format: double
          type: number
    PcCaseDTO:
      type: object
      properties:
        width:
          format: double
          type: number
        length:
          format: double
          type: number
        depth:
          format: double
          type: number
    Person:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        username:
          type: string
        password:
          type: string
        email:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        name:
          type: string
        lastName:
          type: string
        addresses:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Address'
    PersonDTO:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        username:
          type: string
        name:
          type: string
        lastName:
          type: string
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/AddressDTO'
    Processor:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        processorType:
          type: string
        socketType:
          type: string
        numberOfCores:
          format: int32
          type: integer
        numberOfThreads:
          format: int32
          type: integer
        l3Capacity:
          format: int32
          type: integer
    ProcessorDTO:
      type: object
      properties:
        processorType:
          type: string
        socketType:
          type: string
        numberOfCores:
          format: int32
          type: integer
        numberOfThreads:
          format: int32
          type: integer
        l3Capacity:
          format: int32
          type: integer
    Product:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        thumbnail:
          type: string
        pictures:
          type: string
        productCategory:
          $ref: '#/components/schemas/ProductCategory'
        name:
          type: string
        producer:
          type: string
        guarantee:
          type: string
        model:
          type: string
        price:
          format: double
          type: number
        quantity:
          format: int32
          type: integer
        productOrders:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/ProductOrder'
        charger:
          $ref: '#/components/schemas/Charger'
        cooler:
          $ref: '#/components/schemas/Cooler'
        dramMemory:
          $ref: '#/components/schemas/DRAMMemory'
        graphicCard:
          $ref: '#/components/schemas/GraphicCard'
        hardDrive:
          $ref: '#/components/schemas/HardDrive'
        motherboard:
          $ref: '#/components/schemas/Motherboard'
        pcCase:
          $ref: '#/components/schemas/PcCase'
        processor:
          $ref: '#/components/schemas/Processor'
        rate:
          format: float
          type: number
        thumbnailAsByte:
          type: string
        picturesAsBytes:
          type: string
    ProductCategory:
      enum:
      - PROCESSOR
      - GRAPHIC_CARD
      - DRAM_MEMORY
      - MOTHERBOARD
      - COOLER
      - PC_CASE
      - HARD_DRIVE
      - CHARGER
      type: string
...
    ProductOrder:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        quantityProductOrder:
          format: int32
          type: integer
        product:
          $ref: '#/components/schemas/Product'
...
    ProductSearchCriteria:
      type: object
      properties:
        productCategory:
          $ref: '#/components/schemas/ProductCategory'
        name:
          type: string
        producer:
          type: string
        guarantee:
          type: string
        model:
          type: string
        price:
          type: string
        quantity:
          format: int32
          type: integer
    Role:
      enum:
      - ADMIN
      - USER
      - NOT_LOGGED
      type: string
...